Move into custom ray backend
